+++ Packages/Status/Sources/Status/Editor/StatusEditorFocusState.swift
@@ -1,0 +1,5 @@
+import SwiftUI
+
+enum StatusEditorFocusState: Hashable {
+  case main, followUp(index: UUID)
+}
